+++ NEWS.md
@@ -1,6 +1,8 @@
 ## packer 0.0.7.9000
 
 - Ensure existing R files are not overwritten is already existing.
+- Added `make_library` to turn the project into a library 
+callable from outside the bundle.
 
 ## packer 0.0.6